+++ packages/app/src/views/dialog/examples.ts
@@ -1,13 +1,13 @@
 import { autoinject } from 'aurelia-framework';
-import { MdcDialogServiceNew } from '@aurelia-mdc-web/dialog';
+import { MdcDialogService } from '@aurelia-mdc-web/dialog';
 import { ServiceDialog } from './service-dialog';
 import inlineHtml from '!!raw-loader!./inline.html';
 import serviceHtml from '!!raw-loader!./service-dialog.html';
-import serviceCode from '!!raw-loader!./service-dialog';
+import serviceCode from './service-dialog.ts.raw';
 
 @autoinject
 export class Examples {
-  constructor(private dialogService: MdcDialogServiceNew) { }
+  constructor(private dialogService: MdcDialogService) { }
 
   inlineHtml = inlineHtml;
   serviceHtml = serviceHtml;