+++ packages/app/src/views/elevation/examples.ts
@@ -1,5 +1,5 @@
 import basicHtml from '!!raw-loader!./basic.html';
-import basicCode from '!!raw-loader!./basic';
+import basicCode from './basic.ts.raw';
 import basicSass from '!!raw-loader!./basic.scss';
 
 export class Examples {