+++ packages/app/src/views/lookup/examples.ts
@@ -1,13 +1,13 @@
 import { ValidationControllerFactory, ValidationRules, ValidationController, Rule } from 'aurelia-validation';
 import { inject } from 'aurelia-framework';
 import basicHtml from '!!raw-loader!./basic.html';
-import basicCode from '!!raw-loader!./basic';
+import basicCode from './basic.ts.raw';
 import functionHtml from '!!raw-loader!./function.html';
-import functionCode from '!!raw-loader!./function';
+import functionCode from './function.ts.raw';
 import objectsHtml from '!!raw-loader!./objects.html';
-import objectsCode from '!!raw-loader!./objects';
+import objectsCode from './objects.ts.raw';
 import validationHtml from '!!raw-loader!./validation.html';
-import validationCode from '!!raw-loader!./validation';
+import validationCode from './validation.ts.raw';
 
 interface IOption {
   id: number;