+++ packages/app/src/views/segmented-button/examples.ts
@@ -1,7 +1,7 @@
 import multipleHtml from '!!raw-loader!./multiple/multiple.html';
-import multipleCode from '!!raw-loader!./multiple/multiple';
+import multipleCode from './multiple/multiple.ts.raw';
 import singleHtml from '!!raw-loader!./single/single.html';
-import singleCode from '!!raw-loader!./single/single';
+import singleCode from './single/single.ts.raw';
 
 export class Examples {
   multipleHtml = multipleHtml;