+++ packages/menu/src/mdc-menu.ts
@@ -19,11 +19,11 @@ strings.SELECTED_EVENT = strings.SELECTED_EVENT.toLowerCase();
 @customElement('mdc-menu')
 @useView(PLATFORM.moduleName('./mdc-menu.html'))
 export class MdcMenu extends MdcComponent<MDCMenuFoundation> {
-  private menuSurface_: MdcMenuSurface; // assigned in html
+  menuSurface: MdcMenuSurface; // assigned in html
 
   // @child('mdc-list')
   get list_(): MdcList | undefined {
-    const el = this.root.querySelector('mdc-list');
+    const el = this.root.querySelector('mdc-list,mdc-deprecated-list');
     return (el as IMdcListElement)?.au.controller.viewModel;
   }
 
@@ -75,6 +75,17 @@ export class MdcMenu extends MdcComponent<MDCMenuFoundation> {
   @bindable
   maxHeight: number;
 
+  /** Sets whether focus should be restored after the menu is closed */
+  @bindable.booleanAttr
+  skipRestoreFocus: boolean;
+  skipRestoreFocusChanged() {
+    if (this.skipRestoreFocus) {
+      this.root.setAttribute('data-menu-item-skip-restore-focus', 'true');
+    } else {
+      this.root.removeAttribute('data-menu-item-skip-restore-focus');
+    }
+  }
+
   handleKeydown_(evt: KeyboardEvent) {
     this.foundation?.handleKeydown(evt);
     return true;
@@ -91,16 +102,21 @@ export class MdcMenu extends MdcComponent<MDCMenuFoundation> {
   }
 
   get open(): boolean {
-    return this.menuSurface_.open;
+    return this.menuSurface.open;
   }
 
   set open(value: boolean) {
-    this.menuSurface_.open = value;
+    this.menuSurface.open = value;
   }
 
   /** Toggles the menu to open or close */
   toggle() {
     this.open = !this.open;
+  }
+
+  openAnchored(anchor: HTMLElement) {
+    this.anchor = anchor;
+    this.open = true;
   }
 
   get wrapFocus(): boolean {
@@ -150,7 +166,7 @@ export class MdcMenu extends MdcComponent<MDCMenuFoundation> {
   }
 
   async initialise() {
-    await this.menuSurface_.initialised;
+    await this.menuSurface.initialised;
     if (this.defaultFocusState !== undefined) {
       this.defaultFocusStateChanged();
     }
@@ -196,7 +212,7 @@ export class MdcMenu extends MdcComponent<MDCMenuFoundation> {
    * @param corner Default anchor corner alignment of top-left menu corner.
    */
   setAnchorCorner(corner: Corner) {
-    this.menuSurface_.setAnchorCorner(corner);
+    this.menuSurface.setAnchorCorner(corner);
   }
 
   /**
@@ -242,7 +258,7 @@ export class MdcMenu extends MdcComponent<MDCMenuFoundation> {
   }
 
   setAbsolutePosition(x: number, y: number) {
-    this.menuSurface_.setAbsolutePosition(x, y);
+    this.menuSurface.setAbsolutePosition(x, y);
   }
 
   getDefaultFoundation() {
@@ -265,10 +281,14 @@ export class MdcMenu extends MdcComponent<MDCMenuFoundation> {
         const list = this.items;
         list[index].removeAttribute(attr);
       },
+      getAttributeFromElementAtIndex: (index, attr) => {
+        const list = this.items;
+        return list[index].getAttribute(attr);
+      },
       elementContainsClass: (element, className) => element.classList.contains(className),
       closeSurface: (skipRestoreFocus: boolean) => {
         if (!this.stayOpenOnSelection) {
-          this.menuSurface_?.close(skipRestoreFocus);
+          this.menuSurface?.close(skipRestoreFocus);
         }
       },
       getElementIndex: (element) => this.items.indexOf(element),