+++ packages/select/src/mdc-select.ts
@@ -58,7 +58,6 @@ export class MdcSelect extends MdcComponent<MDCSelectFoundationAurelia>{
   private lineRipple?: MdcLineRipple;
   private mdcLabel: MdcFloatingLabel;
   private outline?: MDCNotchedOutline;
-  // private mutationObserver: MutationObserver;
   errors = new Map<IError, boolean>();
 
   /** Sets the select label */
@@ -81,9 +80,9 @@ export class MdcSelect extends MdcComponent<MDCSelectFoundationAurelia>{
   async requiredChanged() {
     await this.initialised;
     if (this.required) {
-      this.selectAnchor.setAttribute('aria-required', 'true');
+      this.selectAnchor?.setAttribute('aria-required', 'true');
     } else {
-      this.selectAnchor.removeAttribute('aria-required');
+      this.selectAnchor?.removeAttribute('aria-required');
     }
     this.foundation?.setRequired(this.required);
     this.taskQueue.queueTask(() => this.foundation?.layout());
@@ -113,12 +112,12 @@ export class MdcSelect extends MdcComponent<MDCSelectFoundationAurelia>{
   @bindable.booleanAttr
   naturalWidth: boolean;
 
-  private initialValue: unknown;
+  private _value: unknown;
   get value(): unknown {
     if (this.foundation) {
       return this.foundation.getValue();
     } else {
-      return this.initialValue;
+      return this._value;
     }
   }
 
@@ -127,11 +126,10 @@ export class MdcSelect extends MdcComponent<MDCSelectFoundationAurelia>{
   }
 
   setValue(value: unknown, skipNotify: boolean = false) {
+    this._value = value;
     if (this.foundation) {
       this.foundation.setValue(value, skipNotify);
       this.foundation.layout();
-    } else {
-      this.initialValue = value;
     }
   }
 
@@ -182,7 +180,7 @@ export class MdcSelect extends MdcComponent<MDCSelectFoundationAurelia>{
 
   initialSyncWithDOM() {
     // set initial value without emitting change events
-    this.foundation?.setValue(this.initialValue, true);
+    this.foundation?.setValue(this._value, true);
     this.foundation?.layout();
     this.errors = new Map<IError, boolean>();
     this.valid = true;