+++ packages/tooltip/src/mdc-default-tooltip-configuration.ts
@@ -1,3 +1,5 @@
 export class MdcDefaultTooltipConfiguration {
+  showDelay?: number;
+  hideDelay?: number;
   scrollHost?: HTMLElement | string;
 }