+++ docs/CHANGELOG.md
@@ -1,3 +1,77 @@
+## [8.0.4](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/compare/v8.0.3...v8.0.4) (2022-03-14)
+
+
+### Bug Fixes
+
+* **select:** reset value when it is not in the new options ([ddb670d](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/commit/ddb670d3f1efc244d8e1624b66a34c47e704e8b5))
+* **tree:** pass filter to child trees ([51f41b1](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/commit/51f41b12f006f507225f2e7808ed7372fca71281))
+
+
+
+## [8.0.3](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/compare/v8.0.2...v8.0.3) (2021-11-17)
+
+
+### Bug Fixes
+
+* **slider:** adjust value to fit into min/max ([65e546c](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/commit/65e546c1b21572144bd8e695f832b9228202bd1a)), closes [#62](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/issues/62)
+
+
+### Features
+
+* **tooltip:** add delay configs ([f4c133c](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/commit/f4c133c418bdd5620b53969229908a45f214c155))
+
+
+
+## [8.0.2](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/compare/v8.0.1...v8.0.2) (2021-11-08)
+
+
+### Bug Fixes
+
+* **base:** do not continue attaching if detached was called ([15df287](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/commit/15df2877f1ba56fef919ee9317f2e89e36981aaf))
+* **checkbox,radio,select,slider,text-field:** consider undefined references ([b67ba9d](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/commit/b67ba9d6c419db18c4036f53f508fb0f5ee7e9af))
+* **menu:** expose menu surface ([40ff16c](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/commit/40ff16c8c565ee82f299fc9c9d724856a47144b5))
+* **text-field:** expose value ([677ddee](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/commit/677ddeef4784a50c51cc5423505bfddf38e8fb4c))
+
+
+
+## [8.0.1](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/compare/v8.0.0...v8.0.1) (2021-10-21)
+
+
+### Bug Fixes
+
+* **lookup:** await for debounce promise on blur ([e574d57](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/commit/e574d574f0b5508d280aa44e58ccd810a1deec37))
+* **lookup:** do not suppress on mouse down ([3d83509](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/commit/3d83509b98d8acebc2c926778c636df2dad04883))
+* **menu-surface:** check for undefined foundation ([4d9b765](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/commit/4d9b7659c64194d943c01b6153e0488953e356b8))
+* **menu:** consider deprecated lists in menus ([22fbd4b](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/commit/22fbd4b704a3809639d3c17a63c50ab09b04ec11))
+
+
+
+# [8.0.0](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/compare/v7.1.4...v8.0.0) (2021-09-27)
+
+### BREAKING CHANGES
+* **dialog:** Old dialog service has been marked as deprecated. See examples for proper DOM.
+
+### Bug Fixes
+
+* **app:** dialog examples ([ffec01b](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/commit/ffec01bdcbf2936622744bbb98479905c54f327f))
+* **chips:** MDC upgrade ([291a417](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/commit/291a4175325a1c1f97683f0fac681008c8cb8752))
+* **list:** handle initial tabindex in item ([e5b522c](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/commit/e5b522cfe3d39bcb89e219bc21dd6b8f938a73c2))
+* **lookup:** include list item into default replaceables ([99e1ce5](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/commit/99e1ce55d76db2648148e9456fbf93eebb960f60))
+* **lookup:** prevent input blur when an item was selected ([5cff8f0](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/commit/5cff8f029401ad79d5aba16595a67d8f0582479a))
+* **slider:** silence TS error ([a7c4794](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/commit/a7c47944884d99e4cabff371e093f00119694eb5))
+* **top-app-bar:** use proper base class ([c473e28](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/commit/c473e289f2abd4b9ad225f42e42e8e02f0f5fd55))
+
+
+### Features
+
+* **data-table:** MDC upgrade ([12f99e7](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/commit/12f99e7bf777baca53a8dc961e82e8befa88afa7))
+* **dialog:** deprecate old dialog service ([42d1a5d](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/commit/42d1a5d9cbf5a4b8a89b70cd8ef850ce3f80d034))
+* **icon-button:** MDC upgrade ([73fc9c8](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/commit/73fc9c86f68ca89bc4eba504a94d5192063edf34))
+* **menu:** add helper method ([039ff29](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/commit/039ff29a07d20a2a15f4ce28646db0ae06d63107))
+* **menu:** MDC upgrade ([f6a244d](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/commit/f6a244d53641abec007e9db5fecd1299a8236d1a))
+
+
+
 ## [7.1.4](https://github.com/aurelia-ui-toolkits/aurelia-mdc-web/compare/v7.1.3...v7.1.4) (2021-09-13)